MenstrualCycle manager by admin and user (not tested)

+++ BE/gender-healthcare-service/src/main/java/com/example/gender_healthcare_service/dto/request/RescheduleBookingRequestDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class RescheduleBookingRequestDTO {
     private LocalDate newBookingDate;
-    private Long newTimeSlotId;
+    private Integer newTimeSlotId;
     private String adminNotes;
 }
